create form, edit smurf file
