+++ smurfs/src/components/App.js
@@ -1,14 +1,14 @@
 import React, { Component } from "react";
 import "./App.css";
 import Smurf from "./Smurf";
-import { createSmurf} from '../components/Form';
+import CreateSmurf from "./SmurfForm";
 class App extends Component {
   render() {
     return (
       <div className="App">
         <h1>SMURFS!</h1>
         <Smurf />
-        <createSmurf />
+        <CreateSmurf />
       </div>
     );
   }
